requirements.txt
flask==0.11
flask-babel==0.11.1
pygal==2.2.2
python-dateutil==2.5.3
wtforms_components==0.10.0
